requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography
requests
